+++ tests/test_plugin_metadata.py
@@ -1,13 +1,13 @@
 import webviz_config
-from webviz_config.plugins import metadata
+from webviz_config.plugins import plugin_project_metadata
 
 
 def test_webviz_config_metadata():
-    meta = metadata["BannerImage"]
 
-    assert meta["dist_name"] == "webviz-config"
-    assert meta["dist_version"] == webviz_config.__version__
+    metadata = plugin_project_metadata["webviz-config"]
 
-    assert meta["documentation_url"] == "https://equinor.github.io/webviz-config"
-    assert meta["download_url"] == "https://pypi.org/project/webviz-config"
-    assert meta["tracker_url"] == "https://github.com/equinor/webviz-config/issues"
+    assert metadata["dist_version"] == webviz_config.__version__
+    assert metadata["documentation_url"] == "https://equinor.github.io/webviz-config"
+    assert metadata["download_url"] == "https://pypi.org/project/webviz-config"
+    assert metadata["source_url"] == "https://github.com/equinor/webviz-config"
+    assert metadata["tracker_url"] == "https://github.com/equinor/webviz-config/issues"